moved bot webhook endpoint from src/main to src/api/routers
